refactor landing page into server component
